Refactor Collect() to make it less nested
This shifts the code left which is often a sign of it being less complex.
It's done by short circuiting early if any of the requests give an error,
removing the for loop to iterate through the reponses and reversing if
statements to have the smaller amount of code inside the block.